Bug 560060 - [Passage] provide "issued licenses" report
 - repackage existing 'customer base' export wizard to a dedicated
package
  - update references
