+++ bundles/org.eclipse.passage.loc.report.ui/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Automatic-Module-Name: org.eclipse.passage.loc.report.ui
 Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.eclipse.passage.loc.report.ui;singleton:=true
-Bundle-Version: 0.1.0.qualifier
+Bundle-Version: 0.2.0.qualifier
 Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
@@ -15,5 +15,5 @@ Require-Bundle: org.eclipse.passage.loc.yars.api;bundle-version="0.1.0";visibili
  org.eclipse.passage.lic.products;bundle-version="0.5.200",
  org.eclipse.osgi,
  org.eclipse.core.runtime
-Export-Package: org.eclipse.passage.loc.report.internal.ui.jface;x-internal:=true
+Export-Package: org.eclipse.passage.loc.report.internal.ui.jface.user
 Bundle-ActivationPolicy: lazy
