+++ bundles/org.eclipse.passage.loc.report.ui/src/org/eclipse/passage/loc/report/internal/ui/jface/user/ExposedExportWizard.java
@@ -10,17 +10,19 @@
  * Contributors:
  *      ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.loc.report.internal.ui.jface;
+package org.eclipse.passage.loc.report.internal.ui.jface.user;
 
 import java.util.function.Consumer;
 
 import org.eclipse.jface.wizard.WizardDialog;
 import org.eclipse.passage.lic.products.registry.ProductRegistry;
-import org.eclipse.passage.loc.report.internal.core.user.CustomerStorage;
 import org.eclipse.passage.loc.report.internal.core.user.CustomerExportService;
+import org.eclipse.passage.loc.report.internal.core.user.CustomerStorage;
 import org.eclipse.swt.widgets.Shell;
 
-@SuppressWarnings("restriction")
+/**
+ * @since 0.2
+ */
 public final class ExposedExportWizard implements Consumer<Shell> {
 
 	private final ProductRegistry products;
